+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Experimental Diffusion Transformer (DiT) model.
 - Improved documentation for diffusion models and diffusion utils.
 - Safe API to override `__init__`'s arguments saved in checkpoint file with
   `Module.from_checkpoint("chkpt.mdlus", override_args=set(...))`.
@@ -35,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Modified `physicsnemo.models.layers.mlp_layers` to support norm, bias, and Conv2D
 - Diffusion utils: `physicsnemo.utils.generative` renamed into `physicsnemo.utils.diffusion`
 - Diffusion models: in CorrDiff model wrappers (`EDMPrecondSuperResolution` and
   `UNet`), the arguments `profile_mode` and `amp_mode` cannot be overriden by